refactor list configuration and untie it from screening configuration so it can be used elsewhere. enhance some ui features for the topics management added the enabled value for section in the api
