+++ packages/app-builder/src/components/ContinuousScreening/shared/RecapRow.tsx
@@ -1,4 +1,4 @@
-import { Capsule } from './Capsule';
+import { Tag } from 'ui-design-system';
 
 export const RecapRow = ({ children }: { children: React.ReactNode }) => {
   return <div className="flex flex-row items-center gap-v2-xs h-[25px]">{children}</div>;
@@ -6,8 +6,11 @@ export const RecapRow = ({ children }: { children: React.ReactNode }) => {
 
 export const RecapCapsule = ({ children }: { children: React.ReactNode }) => {
   return (
-    <Capsule className="bg-grey-background [.group\/recap-valid_&]:bg-surface-card [.group\/recap-valid_&]:dark:border [.group\/recap-valid_&]:dark:border-green-primary">
+    <Tag
+      color="grey"
+      className="bg-grey-background [.group\/recap-valid_&]:bg-surface-card [.group\/recap-valid_&]:dark:border [.group\/recap-valid_&]:dark:border-green-primary"
+    >
       {children}
-    </Capsule>
+    </Tag>
   );
 };
