+++ packages/app-builder/src/components/ListAndTopicConfiguration/index.ts
@@ -1,0 +1,6 @@
+export {
+  ListAndTopicDatasetConfiguration,
+  type ListAndTopicDatasetConfigurationMode,
+} from './context/ListAndTopicDatasetConfiguration';
+export { DatasetSelectionContent } from './DatasetSelectionContent';
+export { getSectionLeafNames } from './dataset-utils';
