+++ packages/marble-api/openapis/marblecore-api/screenings.yml
@@ -445,6 +445,8 @@ components:
                   type: string
                 title:
                   type: string
+        enabled:
+          type: boolean
     OpenSanctionsDatasetFreshnessDto:
       type: object
       required:
